Uniformly use floats in Actor properties
All the underlying implementation and the public entry points have
been switched to floats; the only missing bits are the Actor properties
that deal with positioning and sizing.

This usually means a major pain when dealing with GValues and varargs
functions. While GValue will warn you when dealing with the wrong
conversions, varags will simply die an horrible (and hard to debug)
death via segfault. Nothing much to do here, except warn people in the
release notes and hope for the best.

+++ tests/interactive/test-easing.c
@@ -94,8 +94,8 @@ on_button_press (ClutterActor       *actor,
 
       animation =
         clutter_actor_animate (rectangle, cur_mode, 2000,
-                               "x", (int) event->x,
-                               "y", (int) event->y,
+                               "x", event->x,
+                               "y", event->y,
                                "color", &color,
                                NULL);
     }
